[Reporting] Move "common" types and constants to allow cross-plugin integration
