+++ x-pack/plugins/reporting/common/index.ts
@@ -4,5 +4,14 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
+import { LayoutSelectorDictionary } from './types';
+
 export { CancellationToken } from './cancellation_token';
 export { Poller } from './poller';
+
+export const getDefaultLayoutSelectors = (): LayoutSelectorDictionary => ({
+  screenshot: '[data-shared-items-container]',
+  renderComplete: '[data-shared-item]',
+  itemsCountAttribute: 'data-shared-items-count',
+  timefilterDurationAttribute: 'data-shared-timefilter-duration',
+});
