+++ x-pack/plugins/reporting/public/components/buttons/report_info_button.tsx
@@ -17,8 +17,8 @@ import {
 } from '@elastic/eui';
 import { get } from 'lodash';
 import React, { Component, Fragment } from 'react';
+import { USES_HEADLESS_JOB_TYPES } from '../../../common/constants';
 import { ReportApiJSON } from '../../../common/types';
-import { USES_HEADLESS_JOB_TYPES } from '../../../constants';
 import { ReportingAPIClient } from '../../lib/reporting_api_client';
 
 interface Props {
