+++ x-pack/plugins/reporting/public/components/index.ts
@@ -9,3 +9,4 @@ export { getFailureToast } from './job_failure';
 export { getWarningFormulasToast } from './job_warning_formulas';
 export { getWarningMaxSizeToast } from './job_warning_max_size';
 export { getGeneralErrorToast } from './general_error';
+export { ScreenCapturePanelContent } from './screen_capture_panel_content';
