+++ x-pack/plugins/reporting/public/components/job_download_button.tsx
@@ -7,8 +7,7 @@
 import { EuiButton } from '@elastic/eui';
 import { FormattedMessage } from '@kbn/i18n/react';
 import React from 'react';
-import { JobSummary } from '../';
-import { JobId } from '../../common/types';
+import { JobId, JobSummary } from '../../common/types';
 
 interface Props {
   getUrl: (jobId: JobId) => string;
