+++ x-pack/plugins/reporting/public/index.ts
@@ -5,33 +5,21 @@
  */
 
 import { PluginInitializerContext } from 'src/core/public';
-import { ReportingPublicPlugin } from './plugin';
+import { ScreenCapturePanelContent } from './components/screen_capture_panel_content';
 import * as jobCompletionNotifications from './lib/job_completion_notifications';
-import { JobId, JobStatus } from '../common/types';
+import { ReportingAPIClient } from './lib/reporting_api_client';
+import { ReportingPublicPlugin } from './plugin';
+
+export interface ReportingSetup {
+  components: {
+    ScreenCapturePanel: typeof ScreenCapturePanelContent;
+  };
+}
+
+export type ReportingStart = ReportingSetup;
+
+export { ReportingAPIClient, ReportingPublicPlugin as Plugin, jobCompletionNotifications };
 
 export function plugin(initializerContext: PluginInitializerContext) {
   return new ReportingPublicPlugin(initializerContext);
 }
-
-export { ReportingPublicPlugin as Plugin };
-export { jobCompletionNotifications };
-
-export interface JobSummary {
-  id: JobId;
-  status: JobStatus;
-  title: string;
-  jobtype: string;
-  maxSizeReached?: boolean;
-  csvContainsFormulas?: boolean;
-}
-
-export interface JobSummarySet {
-  completed: JobSummary[];
-  failed: JobSummary[];
-}
-
-type DownloadLink = string;
-export type DownloadReportFn = (jobId: JobId) => DownloadLink;
-
-type ManagementLink = string;
-export type ManagementLinkFn = () => ManagementLink;
