+++ x-pack/plugins/reporting/public/lib/stream_handler.test.ts
@@ -6,8 +6,7 @@
 
 import sinon, { stub } from 'sinon';
 import { NotificationsStart } from 'src/core/public';
-import { JobSummary } from '../';
-import { ReportDocument } from '../../common/types';
+import { JobSummary, ReportDocument } from '../../common/types';
 import { ReportingAPIClient } from './reporting_api_client';
 import { ReportingNotifierStreamHandler } from './stream_handler';
