+++ x-pack/plugins/reporting/server/export_types/csv/index.ts
@@ -5,6 +5,7 @@
  */
 
 import {
+  CSV_JOB_TYPE as jobType,
   LICENSE_TYPE_BASIC,
   LICENSE_TYPE_ENTERPRISE,
   LICENSE_TYPE_GOLD,
@@ -12,7 +13,6 @@ import {
   LICENSE_TYPE_STANDARD,
   LICENSE_TYPE_TRIAL,
 } from '../../../common/constants';
-import { CSV_JOB_TYPE as jobType } from '../../../constants';
 import { CreateJobFn, ExportTypeDefinition, RunTaskFn } from '../../types';
 import { createJobFnFactory } from './create_job';
 import { runTaskFnFactory } from './execute_job';
