+++ x-pack/plugins/reporting/server/lib/layouts/create_layout.ts
@@ -4,8 +4,9 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
+import { LAYOUT_TYPES } from '../../../common/constants';
 import { CaptureConfig } from '../../types';
-import { LayoutInstance, LayoutParams, LayoutTypes } from './';
+import { LayoutInstance, LayoutParams } from './';
 import { PreserveLayout } from './preserve_layout';
 import { PrintLayout } from './print_layout';
 
@@ -13,7 +14,7 @@ export function createLayout(
   captureConfig: CaptureConfig,
   layoutParams?: LayoutParams
 ): LayoutInstance {
-  if (layoutParams && layoutParams.dimensions && layoutParams.id === LayoutTypes.PRESERVE_LAYOUT) {
+  if (layoutParams && layoutParams.dimensions && layoutParams.id === LAYOUT_TYPES.PRESERVE_LAYOUT) {
     return new PreserveLayout(layoutParams.dimensions);
   }
