+++ x-pack/plugins/reporting/server/lib/layouts/preserve_layout.ts
@@ -6,15 +6,10 @@
 
 import path from 'path';
 import { CustomPageSize } from 'pdfmake/interfaces';
-import {
-  getDefaultLayoutSelectors,
-  Layout,
-  LayoutInstance,
-  LayoutSelectorDictionary,
-  LayoutTypes,
-  PageSizeParams,
-  Size,
-} from './';
+import { getDefaultLayoutSelectors } from '../../../common';
+import { LAYOUT_TYPES } from '../../../common/constants';
+import { LayoutSelectorDictionary, PageSizeParams, Size } from '../../../common/types';
+import { Layout, LayoutInstance } from './';
 
 // We use a zoom of two to bump up the resolution of the screenshot a bit.
 const ZOOM: number = 2;
@@ -28,7 +23,7 @@ export class PreserveLayout extends Layout implements LayoutInstance {
   private readonly scaledWidth: number;
 
   constructor(size: Size, layoutSelectors?: LayoutSelectorDictionary) {
-    super(LayoutTypes.PRESERVE_LAYOUT);
+    super(LAYOUT_TYPES.PRESERVE_LAYOUT);
     this.height = size.height;
     this.width = size.width;
     this.scaledHeight = size.height * ZOOM;
