+++ x-pack/plugins/reporting/server/lib/store/index.ts
@@ -4,5 +4,6 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-export { Report, ReportDocument } from './report';
+export { ReportDocument } from '../../../common/types';
+export { Report } from './report';
 export { ReportingStore } from './store';
