+++ x-pack/plugins/reporting/server/lib/tasks/index.ts
@@ -4,8 +4,8 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
+import { ReportSource, TaskRunResult } from '../../../common/types';
 import { BasePayload } from '../../types';
-import { ReportSource } from '../store/report';
 
 /*
  * The document created by Reporting to store as task parameters for Task
@@ -22,11 +22,4 @@ export interface ReportTaskParams<JobPayloadType = BasePayload> {
   meta: ReportSource['meta'];
 }
 
-export interface TaskRunResult {
-  content_type: string | null;
-  content: string | null;
-  csv_contains_formulas?: boolean;
-  size: number;
-  max_size_reached?: boolean;
-  warnings?: string[];
-}
+export { TaskRunResult };
