+++ x-pack/plugins/reporting/server/test_helpers/create_mock_layoutinstance.ts
@@ -4,12 +4,13 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { createLayout, LayoutInstance, LayoutTypes } from '../lib/layouts';
+import { LAYOUT_TYPES } from '../../common/constants';
+import { createLayout, LayoutInstance } from '../lib/layouts';
 import { CaptureConfig } from '../types';
 
 export const createMockLayoutInstance = (captureConfig: CaptureConfig) => {
   const mockLayout = createLayout(captureConfig, {
-    id: LayoutTypes.PRESERVE_LAYOUT,
+    id: LAYOUT_TYPES.PRESERVE_LAYOUT,
     dimensions: { height: 100, width: 100 },
   }) as LayoutInstance;
   mockLayout.selectors = {
